upstream: [move] refactor out jsonrpc lib from move-package-alt (#22399)
