+++ Cargo.toml
@@ -11,6 +11,7 @@ exclude = [
   "external-crates/move/crates/bytecode-verifier-transactional-tests",
   "external-crates/move/crates/enum-compat-util",
   "external-crates/move/crates/invalid-mutations",
+  "external-crates/move/crates/jsonrpc",
   "external-crates/move/crates/language-benchmarks",
   "external-crates/move/crates/module-generation",
   "external-crates/move/crates/move-abstract-interpreter",
