+++ external-crates/move/crates/jsonrpc/src/types.rs
@@ -4,17 +4,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 /// Implementation of the types and protocol for JSON RPC 2.0
-use std::collections::BTreeMap;
-
-use serde::{
-    Deserialize, Deserializer, Serialize, Serializer,
-    de::{DeserializeOwned, Visitor},
-};
-use serde_json::json;
+use serde::{Deserialize, Serialize};
 use thiserror::Error;
-use tokio::io::{
-    AsyncBufRead, AsyncBufReadExt, AsyncRead, AsyncWrite, AsyncWriteExt, BufReader, BufWriter,
-};
 
 pub type RequestID = u64;
 
@@ -25,11 +16,8 @@ pub type BatchResponse<R> = Vec<Response<R>>;
 #[serde(deny_unknown_fields)]
 pub struct Request<A> {
     pub jsonrpc: TwoPointZero,
-
     pub method: String,
-
     pub params: A,
-
     pub id: RequestID,
 }
 
@@ -115,6 +103,8 @@ impl<'de> Deserialize<'de> for TwoPointZero {
 
 #[test]
 fn deserialize() {
+    use serde_json::json;
+
     let value = json!({"result": 0});
     let x: JsonRpcResult<i32> = JsonRpcResult::deserialize(value).expect("foo");
     let JsonRpcResult::Ok { result } = x else {
@@ -123,5 +113,5 @@ fn deserialize() {
     assert_eq!(result, 0);
 
     let v2 = json!({"jsonrpc": "2.0", "id": 0, "result": 0});
-    let response: Response<i32> = Response::deserialize(v2).expect("bar");
+    let _response: Response<i32> = Response::deserialize(v2).expect("bar");
 }
