+++ external-crates/move/crates/move-package-alt/Cargo.toml
@@ -15,6 +15,7 @@ codespan-reporting.workspace = true
 derive-where.workspace = true
 futures.workspace = true
 itertools.workspace = true
+jsonrpc.workspace = true
 move-command-line-common.workspace = true
 move-core-types.workspace = true
 move-package.workspace = true
@@ -57,5 +58,3 @@ path = "src/mocks/mock-resolver-2.rs"
 name = "test_runner"
 harness = false
 
-[[test]]
-name = "jsonrpc_tests"
