+++ external-crates/move/crates/move-package-alt/src/lib.rs
@@ -14,6 +14,5 @@ pub mod errors;
 pub mod flavor;
 pub mod git;
 pub mod graph;
-pub mod jsonrpc;
 pub mod package;
 pub mod schema;
